Fix for Issue #311
    Resolves problem of garbled P25 Phase2 TDMA messages.
    
    There appear to be integer rollovers that occur in the Raspberry Pi environment,
    but not in the X86 environment because:
    
        x86:  __SIZEOF_LONG__ = 8
        rpi4: __SIZEOF_LONG__ = 4
    
    Modifications to lib/lfsr/bit_utils.h, lib/lfsr/lfsr.cxx, and lib/lfsr/lfsr.h
    to use unsigned long long instead of unsigned long resolves the problem for the
    RPi world.
    
    Because
    
        x86:  __SIZEOF_LONG_LONG__ = 8
        rpi4: __SIZEOF_LONG_LONG__ = 8
    
    it doesn't break anything for X86.

+++ lib/lfsr/bit_utils.h
@@ -53,7 +53,7 @@ Eigen::VectorXi* dibits_to_bits(Eigen::VectorXi &dibits) {
 }
 
 // Tested OK
-Eigen::VectorXi *mk_array(unsigned long n, unsigned l) {
+Eigen::VectorXi *mk_array(unsigned long long n, unsigned l) {
 
   Eigen::VectorXi *a = new Eigen::VectorXi(l);
   for(int i=l-1; i>=0; i--) {
@@ -63,11 +63,11 @@ Eigen::VectorXi *mk_array(unsigned long n, unsigned l) {
   return a;
 }
 
-long unsigned mk_int(Eigen::VectorXi a) {
-  long unsigned res = 0L;
+unsigned long long mk_int(Eigen::VectorXi a) {
+  unsigned long long res = 0ULL;
   for (unsigned i=0; i< a.size(); i++) {
-    res *= 2;
-    res += (a(i) & 1);
+    res *= 2ULL;
+    res += (unsigned long long)(a(i) & 1);
   }
   return res;
 }
